refactor(web/main): rename files
